Struct alloc_cortex_m::CortexMHeap

pub struct CortexMHeap { /* fields omitted */ }

Methods

impl CortexMHeap

pub const fn empty() -> CortexMHeap

Crate a new UNINITIALIZED heap allocator

You must initialize this heap using the init method before using the allocator.

pub unsafe fn init(&self, start_addr: usize, size: usize)

Initializes the heap

This function must be called BEFORE you run any code that makes use of the allocator.

start_addr is the address where the heap will be located.

size is the size of the heap in bytes.

Note that:

The heap grows "upwards", towards larger addresses. Thus end_addr must be larger than start_addr

The size of the heap is (end_addr as usize) - (start_addr as usize). The allocator won't use the byte at end_addr.

Unsafety

Obey these or Bad Stuff will happen.

This function must be called exactly ONCE.

size > 0